Update Solar Panel blog posts to App Router Fixes #61
