+++ app/(totbh)/components/AverageConsumedExportedImportedPerHourChart.js
@@ -1,34 +1,57 @@
+"use client";
+
 import { Bar } from "react-chartjs-2";
 import { Chart } from "chart.js/auto";
 import { CategoryScale } from "chart.js/auto";
-import { getDates } from "../lib/getDates";
-
-export default function TotalConsumedExportedPerDayChart({
-  month,
-  year,
-  consumed,
-  exported,
-  imported,
+export default function AverageConsumedExportedImportedPerHour({
+  avgConsumed,
+  avgExported,
+  avgImported,
 }) {
   Chart.register(CategoryScale);
 
   return (
     <Bar
       data={{
-        labels: getDates(parseInt(month), parseInt(year)),
+        labels: [
+          "12am",
+          "1am",
+          "2am",
+          "3am",
+          "4am",
+          "5am",
+          "6am",
+          "7am",
+          "8am",
+          "9am",
+          "10am",
+          "11am",
+          "12pm",
+          "1pm",
+          "2pm",
+          "3pm",
+          "4pm",
+          "5pm",
+          "6pm",
+          "7pm",
+          "8pm",
+          "9pm",
+          "10pm",
+          "11pm",
+        ],
         // datasets is an array of objects where each object represents a set of data to display corresponding to the labels above. for brevity, we'll keep it at one object
         datasets: [
           {
-            data: consumed,
+            data: avgConsumed,
             label: "Consumed (kWh)",
           },
           {
-            data: imported,
             label: "Imported (kWh)",
+            data: avgImported,
           },
           {
             label: "Exported (kWh)",
-            data: exported,
+            data: avgExported,
           },
         ],
       }}
@@ -36,7 +59,7 @@ export default function TotalConsumedExportedPerDayChart({
         plugins: {
           title: {
             display: true,
-            text: "Total Consumed/Imported/Exported Per Day",
+            text: "Average Consumed/Exported/Imported Per Hour",
           },
         },
