+++ app/(totbh)/components/TotalConsumedExportedImportedPerHourChart.js
@@ -1,3 +1,4 @@
+"use client";
 import { Bar } from "react-chartjs-2";
 import { Chart } from "chart.js/auto";
 import { CategoryScale } from "chart.js/auto";
